moved getTime and subfunction to utility.js
